+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- ScriptableObject for DeviceDisplaySet
 - ScriptableObject for DeviceDisplaySettings
 
 ## [1.0.0-preview.1] - 2020-12-30